requirements.txt
langchain
flask
flask-swagger-ui
google.generativeai
openai
streamlit
